Create Basic_Pattern_2.py
Write code to generate a pattern based on the input given. If input 'N' given is 5, the pattern looks like this.
https://s3.amazonaws.com/hr-assets/0/1668079492-5cce0da049-BP_2.png

Input Format

In the first line of the input get the value "N" which determines no. of rows in the pattern.

Constraints

NA

Output Format

Print the pattern

Sample Input 0

2
Sample Output 0

 *
**
Sample Input 1

3
Sample Output 1

  *
 **
***