Create Basic_Pattern_3.py
Noble was asked by the Karunya administration to design a new logo for Karunya. After thinking a lot he came with a logo. If input 'N' is 3 the logo will look like this.
https://s3.amazonaws.com/hr-assets/0/1666838732-a1686ae6c4-BP_3.png

Now he wants you to help him design a program that generates the given logo in all official papers depending on the value of N. Write a program to design the logo.

Input Format

The value of 'N' should be taken as input in the first line.

Constraints

NA

Output Format

The logo should be printed based on the value N

Sample Input 0

2
Sample Output 0

**
*
**
Sample Input 1

3
Sample Output 1

***
**
*
**
***