Create Basic_Pattern_4.py
Write code for generate the given pattern based on the input. The input given by the user determines how many stars are there in the middle row. If the input given is 4, the below given pattern should be printed.
https://s3.amazonaws.com/hr-assets/0/1666854560-dfb59cde15-BP_3.png

Input Format

In the first line of the input get the value "N" which determines the no. of stars in the middle row.

Constraints

NA

Output Format

Print the desired pattern based on the input given by the user.

Sample Input 0

2
Sample Output 0

*
**
*
Sample Input 1

3
Sample Output 1

*
**
***
**
*